+++ src/pages/BetaAccount/ProjectItem.jsx
@@ -1,4 +1,3 @@
-import { gql, useQuery } from "@apollo/client";
 import { formatDistance } from "date-fns";
 import { Link } from "react-router-dom";
 import {
@@ -14,21 +13,8 @@ import {
 import WorkIcon from "@mui/icons-material/Work";
 import DeleteIcon from "@mui/icons-material/Delete";
 
-const GET_PROJECT = gql`
-  query GetProjectIndex($projectId: String!) {
-    projects: beta_get_project(args: { uuid: $projectId }) {
-      updated_at
-      title
-      uname
-      uuid
-    }
-  }
-`;
-
-export const ProjectItem = ({ projectId, showOwner, onDeleteRequest }) => {
-  const { loading, data, error } = useQuery(GET_PROJECT, {
-    variables: { projectId }
-  });
+export const ProjectItem = ({ project, showOwner, onDeleteRequest }) => {
+  const { uuid: projectId, title, uname, updated_at } = project;
 
   const formatDate = (_date) => {
     if (!_date) return "n/a";
@@ -52,18 +38,8 @@ export const ProjectItem = ({ projectId, showOwner, onDeleteRequest }) => {
     onDeleteRequest(projectId);
   };
 
-  // Generic Error
-  if (error)
-    return (
-      <ListItem>
-        <Alert severity="warning" sx={{ flex: 1 }}>
-          {error.message}
-        </Alert>
-      </ListItem>
-    );
-
   // Project Not Found
-  if (!loading && !data.projects.length)
+  if (!updated_at)
     return (
       <ListItem>
         <Alert severity="warning" sx={{ flex: 1 }} onClose={handleDeleteForced}>
@@ -77,7 +53,6 @@ export const ProjectItem = ({ projectId, showOwner, onDeleteRequest }) => {
       to={projectId}
       component={Link}
       secondaryAction={
-        !loading &&
         onDeleteRequest && (
           <IconButton edge="end" onClick={handleDelete}>
             <DeleteIcon />
@@ -91,7 +66,7 @@ export const ProjectItem = ({ projectId, showOwner, onDeleteRequest }) => {
         </Avatar>
       </ListItemAvatar>
       <ListItemText
-        primary={loading ? "..." : data?.projects[0].title}
+        primary={title}
         primaryTypographyProps={{
           color: (theme) =>
             theme.palette.getContrastText(theme.palette.background.default)
@@ -100,11 +75,11 @@ export const ProjectItem = ({ projectId, showOwner, onDeleteRequest }) => {
           <Stack component="span">
             {showOwner && (
               <Typography component="span">
-                shared by: <i>{data?.projects[0].uname}</i>
+                shared by: <i>{uname}</i>
               </Typography>
             )}
             <Typography component="span">
-              last updated: <i>{formatDate(data?.projects[0].updated_at)}</i>
+              last updated: <i>{formatDate(updated_at)}</i>
             </Typography>
           </Stack>
         }